requirements.txt
anthropic>=0.40.0
numpy-financial>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
weasyprint>=62.0
jinja2>=3.1.0
rich>=13.0.0
pytest>=7.0.0
